requirements.txt
Flask==3.1.1
pypdf==5.7.0
openai==1.93.0
PyYAML==6.0.2
mailjet_rest==1.4.0
plotly==5.22.0
PyQt5==5.15.11
python-dotenv==1.1.1
pandas==2.3.0
matplotlib==3.10.3